third_party/blink/web_tests/platform/linux-chrome/external/wpt/preload/prefetch-document-expected.txt

This is a testharness.js-based test.
[FAIL] different-site document prefetch without 'as' should not be consumed
  assert_equals: expected 2 but got 1
[FAIL] different-site document prefetch with 'as=document' should not be consumed
  assert_equals: expected 2 but got 1
Harness: the test ran to completion.
